- Detail Screen VC. - Focus Mapview in those makers or the whole New South Wales state when opening Main Screen.

+++ FloodAreas/FloodAreas/Models/FloodArea.swift
@@ -29,13 +29,13 @@ struct Feature: Codable {
 
 struct Geometry: Codable {
     var type: String?
-    var coordinates: [Double]?
+    var coordinates: [Float]?
     var collections: [Collection]?
 }
 
 struct Collection: Codable {
     var type: String?
-    var coordinates: [Double]?
+    var coordinates: [Float]?
 }
 
 struct Property: Codable {
